Revert "Feature/landing-page"
