+++ src/App.css
@@ -35,63 +35,6 @@
   align-items: center;
 }
 
-.hero-section {
-  height: 100vh; /* Set the height of the hero section to full viewport height */
-  display: flex;
-  justify-content: center; /* Center horizontally */
-  align-items: center; /* Center vertically */
-  text-align: center;
-  background-color: #fafafc; /* Optional: Set background color */
-}
-
-.hero-section h1 {
-  font-size: 3rem;
-  margin-bottom: 1rem;
-}
-
-.hero-section h2 {
-  font-size: 1.5rem;
-  margin-bottom: 2rem;
-}
-
-.hero-image {
-  background-image: url('./cho3.jpg'); /* Replace 'path/to/your/image.jpg' with the actual path to your image */
-  background-size: cover; /* Cover the entire container */
-  background-position: center; /* Center the background image */
-  /* width: 100%;
-  height: 100%;  */
-  border-radius: 10px; /* Optional: Add border radius for rounded corners */
-}
-
-/* .full-width-button {
-  width: 100%;
-} */
-
-
-.hero-image-small {
-  /* background-image: url('cho3.jpg'); */
-  background-color:#c8d1d3;
-  width: 100%;
-  height: 100vh;
-  background-size: cover;
-  /* color: white; */
-  
-}
-
-.hero-image-large {
-  background-image: url('cho3.jpg');
-  background-size: cover;
-  height: 100vh; /* adjust the height according to your needs */
-}
-
-.layout{
-  height: 100vh;
-  justify-content: center;
-  background-color: #fafafc; /* Optional: Set background color */
-
-}
-
-
 @keyframes App-logo-spin {
   from {
     transform: rotate(0deg);
@@ -100,16 +43,3 @@
     transform: rotate(360deg);
   }
 }
-
-
-
-/* Add this CSS to your global styles or a CSS file */
-@keyframes fadeIn {
-  from {
-      opacity: 0;
-  }
-  to {
-      opacity: 1;
-  }
-}
-
