+++ src/components/Button.tsx
@@ -1,43 +1,44 @@
 import React, {ReactNode}  from 'react';
+import  Button  from "@mui/material/Button";
 import  IconButton from '@mui/material/IconButton';
 import Typography from '@mui/material/Typography';
-import Button, { ButtonProps } from '@mui/material/Button';
-import { borderRadius } from '@mui/system';
-import CustomButtonProps from './interfaces/ButtonInterface';
+
+interface CustomButtonProps{
+    onClick: () => void;
+    children: ReactNode;
+    startIcon?: React.ReactNode;
+    endIcon?: React.ReactNode;
+    disabled?:boolean; //disabled prop
+}
 
 
-const CustomButton =({onClick, children, startIcon, endIcon, disabled, width, variant, borderRadius }: CustomButtonProps) => {
+//declare constant 'button' which represents our button component
+const CustomButton =({onClick, children, startIcon, endIcon, disabled }: CustomButtonProps) =>{
 
     return(
-        <Button
-
-            variant={variant || "contained"} 
-            onClick={onClick}
-            className='custom-button'
-            disabled={disabled}
-            sx={{
-                width: width ?? '50%', // Use the width prop or default to 100%
-                borderRadius: borderRadius ?? '16px',
-            }}
-        >
+        <Button variant="contained" onClick={onClick}  className='custom-button' disabled={disabled}>
+           
             {startIcon && (
                 <IconButton>
                     {startIcon}
                 </IconButton>
             )}
+           
+            
+          
+           <Typography>{children}</Typography> 
 
-            <Typography>{children}</Typography>
-
-            {endIcon && (
+           {endIcon && (
                 <IconButton>
                     {endIcon}
                 </IconButton>
             )}
-        </Button>
-    );
-};
+           
 
-export default CustomButton;
+        </Button>
+        );
+        
+};
 
 
 
@@ -69,6 +70,7 @@ export default CustomButton;
 //         Disabled Button without Icon
 //       </CustomButton>  
 
+export default CustomButton;
