--- src/components/Layout.tsx
@@ -1,24 +1,0 @@
-import React from 'react';
-import Header from './Header';
-import Footer from './Footer';
-import theme from '../theme/theme';
-import { ThemeProvider } from '@emotion/react';
-
-
-interface LayoutProps {
-  children: React.ReactNode;
-}
-
-const Layout: React.FC<LayoutProps> = ({ children }) => {
-  return (
-    <ThemeProvider theme={theme}>
-      <Header />
-      <div className='layout'>
-      {children}
-      </div> 
-      <Footer />
-    </ThemeProvider>
-  );
-};
-
-export default Layout;
