--- src/components/interfaces/InputFieldProps.ts
@@ -1,10 +1,0 @@
-export interface Field {
-    id: string;
-    label: string;
-    variant: 'outlined' | 'filled' | 'standard';
-    defaultValue?: string;
-}
-
-export interface Props {
-    fields: Field[];
-}
