--- src/components/interfaces/LayoutProps.ts
@@ -1,7 +1,0 @@
-import { ReactNode } from 'react';
-
-interface LayoutProps{
-    children: ReactNode;
-}
-
-export default LayoutProps;
