+++ src/index.tsx
@@ -4,10 +4,10 @@ import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
+import Landing from './pages/landing';
 import NewCriteria from './pages/newCriteria';
 import OtherNewCriteria from './pages/otherNewCriteria';
 import DecisionModel from './pages/DecisionModel';
-import LandingPage from './pages/Landing';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
@@ -20,7 +20,6 @@ root.render(
         <Route path='/newCriteria' element={<NewCriteria />} />
         <Route path='/othernewCriteria' element={<OtherNewCriteria />} />
         <Route path='/decisionModel' element={<DecisionModel/>} />
-        <Route path='/Landing' element={<LandingPage/>} />
       </Routes>
     </BrowserRouter>
   </React.StrictMode>
