--- src/pages/PreviousDecision.tsx
@@ -1,15 +1,0 @@
-import React from 'react'
-import Search from '../components/Search'
-import Sidebar from '../components/Sidebar'
-
-function PreviousDecision() {
-  return (
-    <div>
-      <Sidebar />
-    </div>
-  )
-}
-
-export default PreviousDecision
- 
-
