+++ src/theme/theme.ts
@@ -1,4 +1,4 @@
-import { createTheme, withTheme } from '@mui/material/styles';
+import { createTheme } from '@mui/material/styles';
 
 const theme =  createTheme({
 	palette: {
@@ -19,45 +19,7 @@ const theme =  createTheme({
 			default: '#fff', 
 		}
 
-	},
-	typography: {
-		h1: {
-		  fontSize: 12,
-		  color: '#071e22',
-		  fontWeight: 'bold',
-		},
-
-		h2: {
-			fontSize: 10,
-			color: '#337357',
-			fontWeight: 200,
-		  },
-
-		h3: {
-			fontSize: 35,
-			color: '#071e22',
-			fontWeight: 300,
-		  },
-
-		subtitle1: {
-			fontSize: 18,
-			color: '#071e22',
-			fontWeight: 500,
-			fontStyle: 'italic',
-		  },
-
-		// body1: {
-		//   fontWeight: 200,
-		//   color: '#071e22',
-		// },
-		button: {
-		  fontStyle: 'italic',
-		  
-		},
-	  },
-
-	
+	}
 });
-
 
 export default theme;
